chore: align wallet tests with test rail
