+++ tests/wallet_main_screen/wallet: footer actions/test_send.py
