+++ tests/wallet_main_screen/wallet: plus button/test_plus_button_manage_account_added_for_new_generated_seed.py
@@ -2,36 +2,38 @@ import time
 
 import allure
 import pytest
-from allure_commons._allure import step
-from . import marks
+from allure import step
+from tests.wallet_main_screen import marks
 
 import constants
 import driver
 from gui.components.signing_phrase_popup import SigningPhrasePopup
-from gui.components.toast_message import ToastMessage
 from gui.components.wallet.authenticate_popup import AuthenticatePopup
+from gui.components.toast_message import ToastMessage
 from gui.main_window import MainWindow
 
 pytestmark = marks
-@allure.testcase('https://ethstatus.testrail.net/index.php?/cases/view/703029', 'Manage a private key imported account')
-@pytest.mark.case(703029)
+
+
+@allure.testcase('https://ethstatus.testrail.net/index.php?/cases/view/703036',
+                 'Manage an account created from the generated seed phrase')
+@pytest.mark.case(703036)
 @pytest.mark.parametrize('user_account', [constants.user.user_account_one])
-@pytest.mark.parametrize('address_pair', [constants.user.private_key_address_pair_1])
 @pytest.mark.parametrize('name, color, emoji, emoji_unicode, '
-                         'new_name, new_color, new_emoji, new_emoji_unicode', [
-                             pytest.param('PrivKeyAcc1', '#2a4af5', 'sunglasses', '1f60e',
-                                          'PrivKeyAcc1edited', '#216266', 'thumbsup', '1f44d')
-                         ])
-@pytest.mark.xfail(reason="https://github.com/status-im/status-desktop/issues/12914")
-def test_plus_button_manage_account_from_private_key(main_screen: MainWindow, user_account, address_pair,
-                                                     name: str, color: str, emoji: str, emoji_unicode: str,
-                                                     new_name: str, new_color: str, new_emoji: str, new_emoji_unicode: str):
-    with step('Import an account within private key'):
+                         'new_name, new_color, new_emoji, new_emoji_unicode, keypair_name', [
+                             pytest.param('SPAcc', '#2a4af5', 'sunglasses', '1f60e',
+                                          'SPAccedited', '#216266', 'thumbsup', '1f44d', 'SPKeyPair')])
+def test_plus_button_manage_account_added_for_new_seed(main_screen: MainWindow, user_account,
+                                                       name: str, color: str, emoji: str,
+                                                       emoji_unicode: str,
+                                                       new_name: str, new_color: str, new_emoji: str,
+                                                       new_emoji_unicode: str,
+                                                       keypair_name: str):
+    with step('Create generated seed phrase wallet account'):
         wallet = main_screen.left_panel.open_wallet()
         SigningPhrasePopup().wait_until_appears().confirm_phrase()
         account_popup = wallet.left_panel.open_add_account_popup()
-        account_popup.set_name(name).set_emoji(emoji).set_color(color).set_origin_private_key(
-            address_pair.private_key).save()
+        account_popup.set_name(name).set_emoji(emoji).set_color(color).set_origin_new_seed_phrase(keypair_name).save()
         AuthenticatePopup().wait_until_appears().authenticate(user_account.password)
         account_popup.wait_until_hidden()
 
@@ -49,15 +51,7 @@ def test_plus_button_manage_account_from_private_key(main_screen: MainWindow, us
             if time.monotonic() - started_at > 15:
                 raise LookupError(f'Account {expected_account} not found in {wallet.left_panel.accounts}')
 
-    with step('Verify that importing private key reveals correct wallet address'):
-        settings_acc_view = (
-            main_screen.left_panel.open_settings().left_panel.open_wallet_settings().open_account_in_settings(name))
-        address = settings_acc_view.get_account_address_value()
-        assert address == address_pair.wallet_address, \
-            f"Recovered account should have address {address_pair.wallet_address}, but has {address}"
-
     with step('Edit wallet account'):
-        main_screen.left_panel.open_wallet()
         account_popup = wallet.left_panel.open_edit_account_popup_from_context_menu(name)
         account_popup.set_name(new_name).set_emoji(new_emoji).set_color(new_color).save()
 
@@ -69,8 +63,8 @@ def test_plus_button_manage_account_from_private_key(main_screen: MainWindow, us
             if time.monotonic() - started_at > 15:
                 raise LookupError(f'Account {expected_account} not found in {wallet.left_panel.accounts}')
 
-    with step('Delete wallet account'):
-        wallet.left_panel.delete_account_from_context_menu(new_name).confirm()
+    with step('Delete wallet account with agreement'):
+        wallet.left_panel.delete_account_from_context_menu(new_name).agree_and_confirm()
 
     with step('Verify toast message notification when removing account'):
         messages = ToastMessage().get_toast_messages
