+++ tests/wallet_main_screen/wallet: plus button/test_plus_button_manage_account_from_private_key.py
@@ -3,7 +3,7 @@ import time
 import allure
 import pytest
 from allure_commons._allure import step
-from . import marks
+from tests.wallet_main_screen import marks
 
 import constants
 import driver
@@ -13,34 +13,24 @@ from gui.components.wallet.authenticate_popup import AuthenticatePopup
 from gui.main_window import MainWindow
 
 pytestmark = marks
-@allure.testcase('https://ethstatus.testrail.net/index.php?/cases/view/703030', 'Manage a seed phrase imported account')
-@pytest.mark.case(703030)
+@allure.testcase('https://ethstatus.testrail.net/index.php?/cases/view/703029', 'Manage a private key imported account')
+@pytest.mark.case(703029)
 @pytest.mark.parametrize('user_account', [constants.user.user_account_one])
+@pytest.mark.parametrize('address_pair', [constants.user.private_key_address_pair_1])
 @pytest.mark.parametrize('name, color, emoji, emoji_unicode, '
-                         'new_name, new_color, new_emoji, new_emoji_unicode, seed_phrase', [
-                             pytest.param('SPAcc24', '#2a4af5', 'sunglasses', '1f60e',
-                                          'SPAcc24edited', '#216266', 'thumbsup', '1f44d',
-                                          'elite dinosaur flavor canoe garbage palace antique dolphin virtual mixed sand '
-                                          'impact solution inmate hair pipe affair cage vote estate gloom lamp robust like'),
-                             pytest.param('SPAcc18', '#2a4af5', 'sunglasses', '1f60e',
-                                          'SPAcc18edited', '#216266', 'thumbsup', '1f44d',
-                                          'kitten tiny cup admit cactus shrug shuffle accident century faith roof plastic '
-                                          'beach police barely vacant sign blossom'),
-                             pytest.param('SPAcc12', '#2a4af5', 'sunglasses', '1f60e',
-                                          'SPAcc12edited', '#216266', 'thumbsup', '1f44d',
-                                          'pelican chief sudden oval media rare swamp elephant lawsuit wheat knife initial')
+                         'new_name, new_color, new_emoji, new_emoji_unicode', [
+                             pytest.param('PrivKeyAcc1', '#2a4af5', 'sunglasses', '1f60e',
+                                          'PrivKeyAcc1edited', '#216266', 'thumbsup', '1f44d')
                          ])
-@pytest.mark.xfail(reason="https://github.com/status-im/status-desktop/issues/12914")
-def test_plus_button_manage_account_from_seed_phrase(main_screen: MainWindow, user_account,
+def test_plus_button_manage_account_from_private_key(main_screen: MainWindow, user_account, address_pair,
                                                      name: str, color: str, emoji: str, emoji_unicode: str,
-                                                     new_name: str, new_color: str, new_emoji: str, new_emoji_unicode: str,
-                                                     seed_phrase: str):
-    with step('Create imported seed phrase wallet account'):
+                                                     new_name: str, new_color: str, new_emoji: str, new_emoji_unicode: str):
+    with step('Import an account within private key'):
         wallet = main_screen.left_panel.open_wallet()
         SigningPhrasePopup().wait_until_appears().confirm_phrase()
         account_popup = wallet.left_panel.open_add_account_popup()
-        account_popup.set_name(name).set_emoji(emoji).set_color(color).set_origin_seed_phrase(
-            seed_phrase.split()).save()
+        account_popup.set_name(name).set_emoji(emoji).set_color(color).set_origin_private_key(
+            address_pair.private_key).save()
         AuthenticatePopup().wait_until_appears().authenticate(user_account.password)
         account_popup.wait_until_hidden()
 
@@ -58,7 +48,15 @@ def test_plus_button_manage_account_from_seed_phrase(main_screen: MainWindow, us
             if time.monotonic() - started_at > 15:
                 raise LookupError(f'Account {expected_account} not found in {wallet.left_panel.accounts}')
 
+    with step('Verify that importing private key reveals correct wallet address'):
+        settings_acc_view = (
+            main_screen.left_panel.open_settings().left_panel.open_wallet_settings().open_account_in_settings(name))
+        address = settings_acc_view.get_account_address_value()
+        assert address == address_pair.wallet_address, \
+            f"Recovered account should have address {address_pair.wallet_address}, but has {address}"
+
     with step('Edit wallet account'):
+        main_screen.left_panel.open_wallet()
         account_popup = wallet.left_panel.open_edit_account_popup_from_context_menu(name)
         account_popup.set_name(new_name).set_emoji(new_emoji).set_color(new_color).save()
 
@@ -70,7 +68,7 @@ def test_plus_button_manage_account_from_seed_phrase(main_screen: MainWindow, us
             if time.monotonic() - started_at > 15:
                 raise LookupError(f'Account {expected_account} not found in {wallet.left_panel.accounts}')
 
-    with step('Delete wallet account with agreement'):
+    with step('Delete wallet account'):
         wallet.left_panel.delete_account_from_context_menu(new_name).agree_and_confirm()
 
     with step('Verify toast message notification when removing account'):
