+++ tests/wallet_main_screen/wallet: plus button/test_plus_button_manage_account_from_seed_phrase.py
@@ -2,35 +2,47 @@ import time
 
 import allure
 import pytest
-from allure import step
-from . import marks
+from allure_commons._allure import step
+from tests.wallet_main_screen import marks
 
 import constants
 import driver
 from gui.components.signing_phrase_popup import SigningPhrasePopup
-from gui.components.wallet.authenticate_popup import AuthenticatePopup
 from gui.components.toast_message import ToastMessage
+from gui.components.wallet.authenticate_popup import AuthenticatePopup
 from gui.main_window import MainWindow
 
 pytestmark = marks
-@allure.testcase('https://ethstatus.testrail.net/index.php?/cases/view/703036',
-                 'Manage an account created from the generated seed phrase')
-@pytest.mark.case(703036)
+
+
+@allure.testcase('https://ethstatus.testrail.net/index.php?/cases/view/703030', 'Manage a seed phrase imported account')
+@pytest.mark.case(703030)
 @pytest.mark.parametrize('user_account', [constants.user.user_account_one])
 @pytest.mark.parametrize('name, color, emoji, emoji_unicode, '
-                         'new_name, new_color, new_emoji, new_emoji_unicode, keypair_name', [
-                             pytest.param('SPAcc', '#2a4af5', 'sunglasses', '1f60e',
-                                          'SPAccedited', '#216266', 'thumbsup', '1f44d', 'SPKeyPair')])
-@pytest.mark.xfail(reason="https://github.com/status-im/status-desktop/issues/12914")
-def test_plus_button_manage_account_added_for_imported_seed_phrase(main_screen: MainWindow, user_account,
-                                                                   name: str, color: str, emoji: str, emoji_unicode: str,
-                                                                   new_name: str, new_color: str, new_emoji: str, new_emoji_unicode: str,
-                                                                   keypair_name: str):
-    with step('Create generated seed phrase wallet account'):
+                         'new_name, new_color, new_emoji, new_emoji_unicode, seed_phrase', [
+                             pytest.param('SPAcc24', '#2a4af5', 'sunglasses', '1f60e',
+                                          'SPAcc24edited', '#216266', 'thumbsup', '1f44d',
+                                          'elite dinosaur flavor canoe garbage palace antique dolphin virtual mixed sand '
+                                          'impact solution inmate hair pipe affair cage vote estate gloom lamp robust like'),
+                             pytest.param('SPAcc18', '#2a4af5', 'sunglasses', '1f60e',
+                                          'SPAcc18edited', '#216266', 'thumbsup', '1f44d',
+                                          'kitten tiny cup admit cactus shrug shuffle accident century faith roof plastic '
+                                          'beach police barely vacant sign blossom'),
+                             pytest.param('SPAcc12', '#2a4af5', 'sunglasses', '1f60e',
+                                          'SPAcc12edited', '#216266', 'thumbsup', '1f44d',
+                                          'pelican chief sudden oval media rare swamp elephant lawsuit wheat knife initial')
+                         ])
+def test_plus_button_manage_account_from_seed_phrase(main_screen: MainWindow, user_account,
+                                                     name: str, color: str, emoji: str, emoji_unicode: str,
+                                                     new_name: str, new_color: str, new_emoji: str,
+                                                     new_emoji_unicode: str,
+                                                     seed_phrase: str):
+    with step('Create imported seed phrase wallet account'):
         wallet = main_screen.left_panel.open_wallet()
         SigningPhrasePopup().wait_until_appears().confirm_phrase()
         account_popup = wallet.left_panel.open_add_account_popup()
-        account_popup.set_name(name).set_emoji(emoji).set_color(color).set_origin_new_seed_phrase(keypair_name).save()
+        account_popup.set_name(name).set_emoji(emoji).set_color(color).set_origin_seed_phrase(
+            seed_phrase.split()).save()
         AuthenticatePopup().wait_until_appears().authenticate(user_account.password)
         account_popup.wait_until_hidden()
