+++ tests/wallet_main_screen/wallet: plus button/test_plus_button_manage_generated_account_custom_derivation_path.py
@@ -3,41 +3,42 @@ import time
 import allure
 import pytest
 from allure_commons._allure import step
-from . import marks
+from tests.wallet_main_screen import marks
 
 import constants
 import driver
 from gui.components.signing_phrase_popup import SigningPhrasePopup
 from gui.components.toast_message import ToastMessage
-from gui.components.wallet.authenticate_popup import AuthenticatePopup
 from gui.main_window import MainWindow
 
 pytestmark = marks
-@allure.testcase('https://ethstatus.testrail.net/index.php?/cases/view/703033', 'Manage a generated account')
-@pytest.mark.case(703033)
+
+
+@allure.testcase('https://ethstatus.testrail.net/index.php?/cases/view/703028', 'Manage a custom generated account')
+@pytest.mark.case(703028)
 @pytest.mark.parametrize('user_account', [constants.user.user_account_one])
-@pytest.mark.parametrize('name, color, emoji, emoji_unicode, '
-                         'new_name, new_color, new_emoji, new_emoji_unicode', [
-                             pytest.param('GenAcc1', '#2a4af5', 'sunglasses', '1f60e',
-                                          'GenAcc1edited', '#216266', 'thumbsup', '1f44d')
-                         ])
-@pytest.mark.xfail(reason="https://github.com/status-im/status-desktop/issues/12914")
-def test_plus_button_manage_generated_account(main_screen: MainWindow, user_account,
-                                              color: str, emoji: str, emoji_unicode: str,
-                                              name: str, new_name: str, new_color: str, new_emoji: str, new_emoji_unicode: str):
+@pytest.mark.parametrize('derivation_path, generated_address_index, name, color, emoji, emoji_unicode, new_name, new_color, new_emoji, new_emoji_unicode',
+                         [
+                            pytest.param('Ethereum', '5', 'Ethereum', '#216266', 'sunglasses', '1f60e', 'EthEdited', '#216266', 'thumbsup', '1f44d'),
+                            pytest.param('Ethereum Testnet (Ropsten)', '10', 'Ethereum Testnet ', '#7140fd', 'sunglasses', '1f60e', 'RopstenEdited', '#216266', 'thumbsup', '1f44d'),
+                            pytest.param('Ethereum (Ledger)', '15', 'Ethereum Ledger', '#2a799b', 'sunglasses', '1f60e', 'LedgerEdited', '#216266', 'thumbsup', '1f44d'),
+                            pytest.param('Ethereum (Ledger Live/KeepKey)', '20', 'Ethereum Ledger Live', '#7140fd', 'sunglasses', '1f60e', 'LiveEdited', '#216266', 'thumbsup', '1f44d'),
+                            pytest.param('N/A', '95', 'Custom path', '#216266', 'sunglasses', '1f60e', 'CustomEdited', '#216266', 'thumbsup', '1f44d')
+])
+def test_plus_button_manage_generated_account_custom_derivation_path(main_screen: MainWindow, user_account,
+                                                                     derivation_path: str, generated_address_index: int,
+                                                                     name: str, color: str, emoji: str,
+                                                                     emoji_unicode: str,
+                                                                     new_name: str, new_color: str, new_emoji: str,
+                                                                     new_emoji_unicode: str):
     with step('Create generated wallet account'):
         wallet = main_screen.left_panel.open_wallet()
         SigningPhrasePopup().wait_until_appears().confirm_phrase()
         account_popup = wallet.left_panel.open_add_account_popup()
-        account_popup.set_name(name).set_emoji(emoji).set_color(color).save()
-        AuthenticatePopup().wait_until_appears().authenticate(user_account.password)
-        account_popup.wait_until_hidden()
+        account_popup.set_name(name).set_emoji(emoji).set_color(color).set_derivation_path(derivation_path,
+                                                                                           generated_address_index,
+                                                                                           user_account.password).save()
 
-    with step('Verify toast message notification when adding account'):
-        assert len(ToastMessage().get_toast_messages) == 1, \
-            f"Multiple toast messages appeared"
-        message = ToastMessage().get_toast_messages[0]
-        assert message == f'"{name}" successfully added'
 
     with step('Verify that the account is correctly displayed in accounts list'):
         expected_account = constants.user.account_list_item(name, color.lower(), emoji_unicode)
@@ -46,6 +47,12 @@ def test_plus_button_manage_generated_account(main_screen: MainWindow, user_acco
             time.sleep(1)
             if time.monotonic() - started_at > 15:
                 raise LookupError(f'Account {expected_account} not found in {wallet.left_panel.accounts}')
+
+    with step('Verify toast message notification when adding account'):
+        assert len(ToastMessage().get_toast_messages) == 1, \
+            f"Multiple toast messages appeared"
+        message = ToastMessage().get_toast_messages[0]
+        assert message == f'"{name}" successfully added'
 
     with step('Edit wallet account'):
         account_popup = wallet.left_panel.open_edit_account_popup_from_context_menu(name)
